Rodrigo

+++ src/principal/IN3.java
@@ -44,7 +44,7 @@ public class IN3 extends Instrumento{
         this.envoltoriaFreqRuido.reset();        
         
         this.ruido = new Ruido(envoltoriaFreqRuido);   
-        this.ruido.setFrequencia ((float) this.frequencia * this.fatorCorte);        
+        this.ruido.setFrequencia (this.frequencia * this.fatorCorte);        
         this.ruido.setFase (this.fase); 
         this.ruido.setDuracao (this.duracao);        
         this.ruido.reset();
